+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/util/FlexGridLayout.java
@@ -1,13 +1,11 @@
 package org.quiltmc.enigma.gui.util;
 
-import com.google.common.base.Preconditions;
 import org.jspecify.annotations.Nullable;
 
 import java.awt.Component;
 import java.awt.Container;
 import java.awt.Dimension;
 import java.awt.LayoutManager2;
-import java.util.Objects;
 
 public class FlexGridLayout implements LayoutManager2 {
 	private final ConstrainedGrid grid = new ConstrainedGrid();
@@ -16,23 +14,23 @@ public class FlexGridLayout implements LayoutManager2 {
 	public void addLayoutComponent(Component component, @Nullable Object constraints) throws IllegalArgumentException {
 		if (constraints == null) {
 			this.addDefaultConstrainedLayoutComponent(component);
-		} else if (constraints instanceof Constraints<?> typedConstraints) {
+		} else if (constraints instanceof FlexGridConstraints<?> typedConstraints) {
 			this.addLayoutComponent(component, typedConstraints);
 		} else {
 			throw new IllegalArgumentException(
 				"constraints type must be %s, but was %s!"
-					.formatted(Constraints.class.getName(), constraints.getClass().getName())
+					.formatted(FlexGridConstraints.class.getName(), constraints.getClass().getName())
 			);
 		}
 	}
 
-	public void addLayoutComponent(Component component, @Nullable Constraints<?> constraints) {
+	public void addLayoutComponent(Component component, @Nullable FlexGridConstraints<?> constraints) {
 		if (constraints == null) {
 			this.addDefaultConstrainedLayoutComponent(component);
 		} else {
 			final int x;
 			final int y;
-			if (constraints instanceof Constraints.Absolute absolute) {
+			if (constraints instanceof FlexGridConstraints.Absolute absolute) {
 				x = absolute.x;
 				y = absolute.y;
 			} else {
@@ -54,20 +52,17 @@ public class FlexGridLayout implements LayoutManager2 {
 	}
 
 	private int getRelativeX() {
-		return this.grid.isEmpty() ? Constraints.Absolute.DEFAULT_X : this.grid.getMaxXOrThrow() + 1;
+		return this.grid.isEmpty() ? FlexGridConstraints.Absolute.DEFAULT_X : this.grid.getMaxXOrThrow() + 1;
 	}
 
 	private int getRelativeY() {
-		return this.grid.isEmpty() ? Constraints.Absolute.DEFAULT_Y : this.grid.getMaxYOrThrow();
+		return this.grid.isEmpty() ? FlexGridConstraints.Absolute.DEFAULT_Y : this.grid.getMaxYOrThrow();
 	}
 
 	@Override
 	public void removeLayoutComponent(Component component) {
 		this.grid.remove(component);
 	}
-
-	@Override
-	public Dimension maximumLayoutSize(Container target) { }
 
 	@Override
 	public float getLayoutAlignmentX(Container target) {
@@ -85,6 +80,9 @@ public class FlexGridLayout implements LayoutManager2 {
 	}
 
 	@Override
+	public Dimension maximumLayoutSize(Container target) { }
+
+	@Override
 	public Dimension preferredLayoutSize(Container parent) { }
 
 	@Override
@@ -99,274 +97,6 @@ public class FlexGridLayout implements LayoutManager2 {
 		BEGIN, CENTER, END
 	}
 
-	public static sealed abstract class Constraints<C extends Constraints<C>> {
-		private static final int DEFAULT_PRIORITY = 0;
-		private static final int DEFAULT_WIDTH = 1;
-		private static final int DEFAULT_HEIGHT = 1;
-		private static final boolean DEFAULT_FILL_X = false;
-		private static final boolean DEFAULT_FILL_Y = false;
-		private static final Alignment DEFAULT_X_ALIGNMENT = Alignment.BEGIN;
-		private static final Alignment DEFAULT_Y_ALIGNMENT = Alignment.CENTER;
-
-		public static Relative createRelative() {
-			return Relative.of();
-		}
-
-		public static Absolute createAbsolute() {
-			return Absolute.of();
-		}
-
-		private static Alignment requireNonNullAlignment(Alignment alignment) {
-			return Objects.requireNonNull(alignment, "alignment must not be null!");
-		}
-
-		int width;
-		int height;
-
-		boolean fillX;
-		boolean fillY;
-
-		Alignment xAlignment;
-		Alignment yAlignment;
-
-		int priority;
-
-		private Constraints(
-			int width, int height,
-			boolean fillX, boolean fillY,
-			Alignment xAlignment, Alignment yAlignment,
-			int priority
-		) {
-			this.width = width;
-			this.height = height;
-
-			this.fillX = fillX;
-			this.fillY = fillY;
-
-			this.xAlignment = xAlignment;
-			this.yAlignment = yAlignment;
-
-			this.priority = priority;
-		}
-
-		public C width(int width) {
-			Preconditions.checkArgument(width > 0, "width must be positive!");
-			this.width = width;
-			return this.getSelf();
-		}
-
-		public C height(int height) {
-			Preconditions.checkArgument(height > 0, "height must be positive!");
-			this.height = height;
-			return this.getSelf();
-		}
-
-		public C fillX(boolean fill) {
-			this.fillX = fill;
-			return this.getSelf();
-		}
-
-		public C fillX() {
-			return this.fillX(true);
-		}
-
-		public C fillY(boolean fill) {
-			this.fillY = fill;
-			return this.getSelf();
-		}
-
-		public C fillY() {
-			return this.fillY(true);
-		}
-
-		public C fill(boolean x, boolean y) {
-			this.fillX(x);
-			this.fillY(y);
-			return this.getSelf();
-		}
-
-		public C fillBoth() {
-			return this.fill(true, true);
-		}
-
-		public C xAlignment(Alignment alignment) {
-			this.xAlignment = requireNonNullAlignment(alignment);
-			return this.getSelf();
-		}
-
-		public C alignLeft() {
-			return this.xAlignment(Alignment.BEGIN);
-		}
-
-		public C alignRight() {
-			return this.xAlignment(Alignment.END);
-		}
-
-		public C yAlignment(Alignment alignment) {
-			this.yAlignment = requireNonNullAlignment(alignment);
-			return this.getSelf();
-		}
-
-		public C alignTop() {
-			return this.yAlignment(Alignment.BEGIN);
-		}
-
-		public C alignBottom() {
-			return this.yAlignment(Alignment.END);
-		}
-
-		public C align(Alignment x, Alignment y) {
-			this.xAlignment(x);
-			this.yAlignment(y);
-			return this.getSelf();
-		}
-
-		public C alignCenter() {
-			return this.align(Alignment.CENTER, Alignment.CENTER);
-		}
-
-		public C priority(int priority) {
-			this.priority = priority;
-			return this.getSelf();
-		}
-
-		public abstract C copy();
-
-		protected abstract C getSelf();
-
-		public static final class Relative extends Constraints<Relative> {
-			public static Relative of() {
-				return new Relative(
-					DEFAULT_WIDTH, DEFAULT_HEIGHT,
-					DEFAULT_FILL_X, DEFAULT_FILL_Y,
-					DEFAULT_X_ALIGNMENT, DEFAULT_Y_ALIGNMENT,
-					DEFAULT_PRIORITY
-				);
-			}
-
-			private Relative(
-				int width, int height,
-				boolean fillX, boolean fillY,
-				Alignment xAlignment, Alignment yAlignment,
-				int priority
-			) {
-				super(width, height, fillX, fillY, xAlignment, yAlignment, priority);
-			}
-
-			@Override
-			public Relative copy() {
-				return new Relative(
-					this.width, this.height,
-					this.fillX, this.fillY,
-					this.xAlignment, this.yAlignment,
-					this.priority
-				);
-			}
-
-			public Absolute toAbsolute() {
-				return new Absolute(
-					Absolute.DEFAULT_X, Absolute.DEFAULT_Y,
-					this.width, this.height,
-					this.fillX, this.fillY,
-					this.xAlignment, this.yAlignment,
-					this.priority
-				);
-			}
-
-			public Absolute toAbsolute(int x, int y) {
-				return this.toAbsolute().pos(x, y);
-			}
-
-			@Override
-			protected Relative getSelf() {
-				return this;
-			}
-		}
-
-		public static final class Absolute extends Constraints<Absolute> {
-			private static final int DEFAULT_X = 0;
-			private static final int DEFAULT_Y = 0;
-
-			public static Absolute of() {
-				return new Absolute(
-					DEFAULT_X, DEFAULT_Y,
-					DEFAULT_WIDTH, DEFAULT_HEIGHT,
-					DEFAULT_FILL_X, DEFAULT_FILL_Y,
-					DEFAULT_X_ALIGNMENT, DEFAULT_Y_ALIGNMENT,
-					DEFAULT_PRIORITY
-				);
-			}
-
-			private int x;
-			private int y;
-
-			private Absolute(
-				int x, int y,
-				int width, int height,
-				boolean fillX, boolean fillY,
-				Alignment xAlignment, Alignment yAlignment,
-				int priority
-			) {
-				super(width, height, fillX, fillY, xAlignment, yAlignment, priority);
-
-				this.x = x;
-				this.y = y;
-			}
-
-			public Absolute x(int x) {
-				this.x = x;
-				return this;
-			}
-
-			public Absolute nextRow() {
-				this.x++;
-				this.y = 0;
-				return this;
-			}
-
-			public Absolute y(int y) {
-				this.y = y;
-				return this;
-			}
-
-			public Absolute nextColumn() {
-				this.y++;
-				return this;
-			}
-
-			public Absolute pos(int x, int y) {
-				this.x(x);
-				this.y(y);
-				return this;
-			}
-
-			@Override
-			public Absolute copy() {
-				return new Absolute(
-					this.x, this.y,
-					this.width, this.height,
-					this.fillX, this.fillY,
-					this.xAlignment, this.yAlignment,
-					this.priority
-				);
-			}
-
-			public Relative toRelative() {
-				return new Relative(
-					this.width, this.height,
-					this.fillX, this.fillY,
-					this.xAlignment, this.yAlignment,
-					this.priority
-				);
-			}
-
-			@Override
-			protected Absolute getSelf() {
-				return this;
-			}
-		}
-	}
-
 	record Constrained(
 		Component component, int width, int height, boolean fillX, boolean fillY, Alignment xAlignment,
 		Alignment yAlignment, int priority
@@ -374,14 +104,14 @@ public class FlexGridLayout implements LayoutManager2 {
 		static Constrained defaultOf(Component component) {
 			return new Constrained(
 				component,
-				Constraints.DEFAULT_WIDTH, Constraints.DEFAULT_HEIGHT,
-				Constraints.DEFAULT_FILL_X, Constraints.DEFAULT_FILL_Y,
-				Constraints.DEFAULT_X_ALIGNMENT, Constraints.DEFAULT_Y_ALIGNMENT,
-				Constraints.DEFAULT_PRIORITY
+				FlexGridConstraints.DEFAULT_WIDTH, FlexGridConstraints.DEFAULT_HEIGHT,
+				FlexGridConstraints.DEFAULT_FILL_X, FlexGridConstraints.DEFAULT_FILL_Y,
+				FlexGridConstraints.DEFAULT_X_ALIGNMENT, FlexGridConstraints.DEFAULT_Y_ALIGNMENT,
+				FlexGridConstraints.DEFAULT_PRIORITY
 			);
 		}
 
-		Constrained(Component component, Constraints<?> constraints) {
+		Constrained(Component component, FlexGridConstraints<?> constraints) {
 			this(
 				component,
 				constraints.width, constraints.height,
